Move experiment generation code to cutting_experiments.py
